requirements.txt
avro
msgpack-python
scikit-learn
py-ubjson
pandas
treelite
treelite-runtime
xgboost